+++ lib/webmock/test_unit.rb
@@ -1,6 +1,8 @@
 require 'test/unit'
 require 'webmock'
 
+WebMock.enable!
+
 module Test
   module Unit
     class TestCase