+++ spec/acceptance/excon/excon_spec_helper.rb
@@ -5,14 +5,20 @@ module ExconSpecHelper
   def http_request(method, uri, options = {}, &block)
     Excon.defaults[:ssl_verify_peer] = false
     uri      = Addressable::URI.heuristic_parse(uri)
-    uri      = uri.omit(:userinfo).to_s.gsub(' ', '%20')
+    uri      = uri.to_s.gsub(' ', '%20')
+
+    excon_options = {}
+
+    if basic_auth = options.delete(:basic_auth)
+      excon_options = {user: basic_auth[0], password: basic_auth[1]}
+    end
 
     if Gem::Version.new(Excon::VERSION) < Gem::Version.new("0.29.0")
       options  = options.merge(:method => method, :nonblock => false) # Dup and merge
-      response = Excon.new(uri).request(options, &block)
+      response = Excon.new(uri, excon_options).request(options, &block)
     else
       options  = options.merge(:method => method) # Dup and merge
-      response = Excon.new(uri, :nonblock => false).request(options, &block)
+      response = Excon.new(uri, excon_options.merge(nonblock: false)).request(options, &block)
     end
 
     headers  = WebMock::Util::Headers.normalize_headers(response.headers)